change: cleanup and changelog
