+++ CHANGELOG.mdx
@@ -17,6 +17,10 @@ Prefix the change with one of these keywords:
 
 ## [Unreleased]
 
+### Changed
+
+- Image slider subcomponent structure, reworked padding to prevent cut off
+
 ## [0.18.13]
 
 ### Fixed
